socketserver/requirements.txt
# WebSocket Server Requirements

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=14.0

# Google Cloud
google-cloud-aiplatform>=1.122.0
vertexai

# Utilities
python-dotenv>=1.0.0
